requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.7.4
decorator==5.1.1
Django==5.1
django-elasticsearch-dsl==8.0
django-nine==0.2.7
django-rest-framework==0.1.0
djangorestframework==3.15.2
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.1
executing==2.0.1
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==1.26.19
wcwidth==0.2.13
